refactor: post 패키지의 테스트 코드 정리
